+++ src/tests/e2e/mod.rs
@@ -1,0 +1,2 @@
+pub mod cases;
+mod remote_runner;
