+++ src/tests/integration/mod.rs
@@ -1,4 +1,3 @@
-pub mod e2e;
 pub mod basic;
 pub mod close_pane;
 pub mod compatibility;
