+++ src/tests/mod.rs
@@ -1,5 +1,6 @@
 pub mod fakes;
 pub mod integration;
+pub mod e2e;
 pub mod possible_tty_inputs;
 pub mod tty_inputs;
 pub mod utils;
